changes to support large objects
git-svn-id: https://atmos-c.googlecode.com/svn/trunk@21 cb61ccbc-2dfe-265f-a3f4-a677fceac7d0

+++ atmos_rest.h
@@ -38,7 +38,7 @@ typedef struct System_meta {
     char gid[GIDSIZE];
     char objectid[OBJECTIDSIZE];
     char objname[1024];
-    int size;
+    unsigned long long size;
     int nlink;
     char policyname[POLICYSIZE];
 } system_meta;
